+++ wcivf/apps/elections/views/postcode_view.py
@@ -6,8 +6,9 @@ from django.conf import settings
 from django.http import HttpResponse, HttpResponseRedirect
 from django.utils import timezone
 from django.views.generic import TemplateView, View
+from elections.devs_dc_client import InvalidPostcodeError
 from elections.dummy_models import DummyPostElection, dummy_polling_station
-from elections.models import LOCAL_TZ, InvalidPostcodeError
+from elections.models import LOCAL_TZ
 from icalendar import Calendar, Event, vText
 from parishes.models import ParishCouncilElection
